Update all headers to match upstream version at time of import
When the headers were imported into the linux tree, the copyright
header was removed and replaced with an SPDX liscense identifier
instead.

The corresponding license texts have been imported from the linux
repo into the LICENSES folder.

There was also a single whitespace change in types.h to replace
an errant tab with a space.

Signed-off-by: Charles Mirabile <cmirabil@redhat.com>